Lots of work to be done, but we'll see how it goes.
In its current state, the WWF solver is heavily based off of a program I recently finished for finding the best possible hand in Scrabble, and wouldn't do anything with the board at all. Previous installments of that program involved obtaining the list of letters from the hand, then scrambling them until words were found which was a horrible system. After implementing the much better system of using the hand to navigate a binary tree system for all my available words, speed of searching for words was reduced from ~1.5 seconds for a hand of 7 to milliseconds for a hand of 7. This massive optimization led me to want to build a system that would take the entire board into consideration instead of just the hand.

This should be an all encompassing solver that will consider all possibilities for the best possible play using legal moves that leads to the most points. The only current additional files it uses are the ENABLE list of 172,000+ English words (which WWF is based off of, but they added some of their own, roughly 1000 words) and a .txt file that represents the board. I didn't really want to put the board in the code because that would take up a big unsightly block of code, and leaving it in a configurable .txt file makes it easier to change if the board ever changes, or if you want to change it to use the Scrabble board.

Todo:
- Fully implement board import and use in recursive search function
- Change search function to better work with being forced to use letters from board, and to use another function to check current tile placement to be sure all words stemming from it are legal.
- Create base function that will launch the recursive search from coordinates and in directions around the board.
- Create function to score all changed/new words in a way that accurately reflects WWF scoring

- Consider adding in my additional word bank as a set of words that might be accepted, but also might fail word checks.
- (Long term) Consider adding in graphics for better display instead of only using the python console.