Some major usability changes.
Added in quite a few functions that are building up to the final product. Functions to check words in the board, create the list of potential beginning of words, etc. Considering a few smaller systematic changes that could make working with certain parts of creating words on the board easier, but still got a few details to sort out first.

Still got a ways to go, but the first working version should be soon-ish.