It lives!
Program is completely functional. It uses "wwf board.txt" as a map for tile and word multipliers, takes board.txt as an input for the current state of the board, and board.txt can be deleted to quickly reset the board back to blank spaces of the same size as "wwf board.txt". It does however require that "board.txt" be of the same dimensions as "wwf board.txt" otherwise it'll run into problems being unsure of how to handle additional/lacking information. Probably going to do something about that later.

Besides that, it'll find all possible locations for words to be placed, find all words that work in said locations, assess their point values (including bingo bonus for using all letters of default 35 points) and then rank all valid options by their score. After displaying scores sorted in descending order of their point values, it will display a few of the best options. Changing the result_limit value to 0 will have it always display all results. This defaults to 5, because my test example given hand asdfqwer yields 2500 results, which I think is too much to want to sort through. It defaults to a value of 5 because it isn't showing the highest words, but the highest scores and all words that yield such scores. As such, even showing 5 can still display quite a lot of words, if there isn't a lot of deviation in score values among the highest ranking words (possible if the same word is able to be placed in different locations for the same score).

Besides that, I removed an unnecessary function, moved around some constants, and added in basic and simple user interface that just refreshes the board after each run before entering in the next set of letters. Glad it's finally done and in a working state. It also calculates the whole thing pretty quickly. My test example took 0.7 seconds to finish calculating its 2500 results on a decent processor. I'll take a look at what's left on my original Todo list to see what remains. For now, I'm just glad it's in a fully functional state.