Wait, no, now it lives.
Messed up on mixing up two of my functions. I apparently in all my testing hadn't run into this situation where the recursive search attempts to extend beyond the bottom of the board, which is one of the two things check_coords() is for (checking coordinates beyond the far edge and the close edge to prevent IndexError or overlapping on itself respectively).

Should work for all boards now. I'll test more with "real-world" examples later when I'm not coding at 4am.