Bugfix(es)!
Made two kinda major bugfixes.

(1) During score assessment of words, it first scans the board to find all words, then yield all words that have been changed by placed letters. In the function, this is handled by new = new or d['new'] which essentially means for each letter, it checks if that letter is new. If any of the letters have new set to True, then the whole thing will return True, because or'ing across the whole word. The bug that arose from this was that the new variable didn't properly reset after checking one word. Which isn't a concern, usually, but it had a butterfly effect to causing very different results in possible situations. The one situation it wouldn't properly reset the value of new, was during the for loop of different directions to check for a word on a given y, x coordinate. It first checks the potential word to the right, then the potential word down below the given y, x.

If a tile was the first of two words leading down and right, and if the right-facing word was marked as new, then the downward facing word would also be marked as new. This lead to that word being passed along to the score assessment, which in my situation lead to the program reporting a single tile could net 21 points, instead of 9. Kinda niche situation to run across, but with how complex WWF boards can get, it's very possible and can make low-value words return much higher scores than they actually yield.

(2) Import board not importing word and tile multiplier values. Rather it did actually import them perfectly fine, however the last for loops that were meant to remove multipliers from tiles already used ran for every single tile on the board. An 'if let:' statement was changed to 'if let != blank_space:' and now it works just fine.

Still looking for more bugs so I can get it to perfectly imitate WWF rules and scoring, and once I'm confident it's working just like that, I'll go back to the Todo list. After having had to input a few boards, I'm really starting to see the appeal of having a GUI to make it easier to see what I'm doing while entering in tiles.