Fixed more scoring issues.
As it turns out, the original source I had for point values for letters in WWF was wrong (I don't play the game myself, so I checked a website, not the app) and that had to be updated. I also found another bug with how it handles the bingo bonus. The previous setup only cared if you used all your letters, which was a nice enough system if you don't want your system to be based around constants. As it turns out, however, that is not the ruling in WWF. You must use all 7 letters in your hand, so I had to change out how it handles that from "if used all letters" to "if initial hand length minus number of letters used is greater than number of letters required for bonus." And lastly removed a bit of test code at the end that I forgot to remove before the last commit.