A few general fixes, some major, some minor.
In order of changes in files, added 'qi' to the enable1.txt dictionary, as that's a pretty important word that I happen to know WWF accepts. I'm not planning on adding every word that I find that WWF accepts, but rather just words that are instrumental. And this has been the single most important word I've found that enable1 doesn't include.

Besides that, changed a few comments to be a little more descriptive, made print_board() look a little better (has a nicer border, shows the grid and multiplier tiles), moved around a function (was for ease of testing purposes), fixed a bug where the solver would reuse letters in the hand (from using a global hand variable instead of its new local hand_sect variable), added in a function to reduce the number of potentials (more on this later), added in support for when the board is blank (used to require at least one letter to generate potentials around), and probably most importantly, fixed a relatively major oversight, where required length of a word wasn't taken into account (more on this later, again). And lastly, added in a simple display to show how many tiles need to be processed in the recursive search, and how far along the program is with all those tiles.

The recursive search function was originally intended to include the required length (which is why there was always a parameter for it), but while creating the rest of the program, it slipped my mind to add it in. Because this is a part of the program that's deeply rooted into the potential generation, several function parameters had to be changed to account for this. Most importantly, a simple method for eliminating repeat potentials was no longer possible. This was the reason for adding in the reduce_potential_list() function.

The reduce_potential_list() function was designed to run between n+1 and 2*n iterations (in practice, roughly 1.25*n). It first generates a set of nested dictionaries all containing y, x, and directional data in the same style as my recursive tree function, and then finally containing a list of all distance values. Because the recursive search will also generate as long as possible words that it can generate for a given tile, it only needs to know the bare minimum length, and can get rid of all additional data. Because of this, it can then run through all the nested dictionaries, and append to the output_list tile data, and the minimum distance required. I haven't yet done a lot of testing, but of the tests I've done for this code, I've found that it tends to reduce the number of potential tiles by 75%.

The major net gain for adding in this distance fix and optimization is so that all proposed moves are indeed legal, as without the required length parameter, the program will think it can place words willy-nilly completely detached from the rest of the letters on the board. Now it knows better than that. Will update more as I find more issues needing fixing, or if I make any progress with a proper GUI.