Fixed wildcard scoring.
Minor fix for wildcards, where they could count as the point values of whatever tile they were pretending to be. Fixed this, by adding in a wildcard parameter, that will change the tile multiplier of a wildcard tile to 0. I wouldn't mind implementing a different fix, like one that better acknowledges wildcard tiles after they've been placed on the board, or for when their placement on the board is part of the given state of the board, but that would require changes to the entire design of the system. And at that, I haven't yet found a situation in WWF that this is significant. I'm sure there's plenty of situations, but I haven't yet run into one.

For now, the easy fix. I'll see about fixing the rest of the issue soon.