Added in definite/possible words, and a few minor changes/bugfixes.
Because the program is using the same list of words that WWF's own list of words is based off of, every word that is returned from enable1 list is definitely a word that can be used. However I do know that they've added many words, and I don't have access to their list of words. I do however have a list of 11,000 words that are not in enable1, that are accepted by the 1998 Scrabble Dictionary, and/or by Microsoft's Wordament game (a game very similar to Boggle). I don't however know whether or not any of these are accepted by WWF. So, I've added in a system that will display to the user all possible words on the board that definitely can be used, and another set of words that might be able to be used. I tried to make this obvious in how it presents the words, but I don't think it is quite yet.

Besides that, fixed a bug that allowed repeat results to be passed to display. As it turns out, when trying to find repeat results, it's more effective to sort by the word than by score. While it was sorting by score, under very specific situations it would not sort like-scored results by their name. My method for finding repeat results absolutely requires everything to be sorted well, so when things aren't sorted quite right, it can't find repeats effectively. Instead, it now sorts by the result word, then after omitting all repeats, sorts by score before returning the result_list.

Changed how some details for print_board. Instead of using giant blocks for tile multiplier spots, it now uses smaller dots, then larger dots for word multiplier spots. Also added in a toggle for whether or not it displays the multipliers. It did help when the multipliers were giant blocks, and I guess it isn't necessary now since the multiplier is much smaller, but eh whatever. Leaving it in for yay options.

Besides those, all remaining changes are relevant to adding in the extra possible results, and to minor changes to how it displays results.

Lastly, I'm now very seriously considering adding in a GUI based on pygame. Going to see if I can make a mockup and work on some assets soon before making the most basic versions.