Fixed a pretty big bug, and changed out Definite/Possible for Certain/Uncertain.
That's not just a cosmetic change either. The old system of Definite/Possible used two separate lists of words and ran the same recursive function twice, one for each list of words. They were arbitrarily different, and one was simply labeled for words it knew were definitely accepted by WWF. The other was used as words that could possibly be accepted by WWF. I found this issue while attempting to add in a new word that I knew definitely should show up in a particular situation. I no longer remember the word I was trying to add, but point still being. I was testing a particular board configuration, and we knew that one word should show up, as that was our current best guess. When I tested that word in my solver, it didn't show up, so I added it to enable2, and then looked for it again. For some reason, it didn't show up. I knew it was in the list of words, I could test and find that it was indeed in the enable2_tree, and it would show up when tested, but nonetheless, this word wouldn't be returned as a possible result. After debugging with print all over various functions, I noticed that literally this word sometimes would be returned as a valid possible word, and sometimes it wouldn't.

It finally hit me that because the previous system was using two separate distinct lists of words, that newly added word was in fact sometimes a word, and sometimes not. More importantly, all perpendicular and otherwise newly created words were also sometimes valid and sometimes not valid. The reason this new word couldn't be added in was because there was another valid, definite word that was required. And because definite words weren't in the list for possible words, that definite word didn't return true when checking the perpendicular words. An example of this would be trying to lay down the (previously) "possible" word ABAYA, and having the final A rest between two D's. This could potentially be a decent combination of making the word ABAYA and DAD simultaneously. The previous system would find ABAYA in the word_tree just fine, but when checking the perpendicular DAD, wouldn't find it in its list. This would be because DAD is a word that's definitely known to be accepted by WWF, and thus is not in the possible word list. Therefore, ABAYA would not work, because it requires invalid words.

Because the two sets of words were completely independent of each other, when one actually did require the other, they wouldn't and couldn't work together.

I only knew of two ways to fix this issue. One was my temporary fix of add all the contents of enable1.txt to enable2.txt (which worked for confirming that particular word was in fact the best choice), and the second was the have both sets of words in the same list (and tree) but differentiate which ones are definite and possible. The problem with the first fix of course being that it's running the exact same search all over again with more words, and returning more of the exact same results, with the sprinkled in possibility of different words. Not a very good fix. I also took the opportunity to try and change the word choice, as I didn't particularly care for definite/possible. I believe Certain/Uncertain better conveys the intent of the distinction, and the state of the result. The new Certain/Uncertain words are now both contained within the same result list, and are displayed as if a word is Certain or Uncertain. A word marked as Certain means that it itself is certain to work, as well as all other newly created words from it (perpendicular, adjacent, etc.) are also certain to work. Uncertain words on the other hand may be uncertain because the word itself is marked as uncertain (previous example of ABAYA, an uncertain word, creating DAD, a certain word), or because newly created words are uncertain (example: laying down APPLICATION, and one of the P's completes the Uncertain word APP).

This new system should find more results as it's better using all available words, and will have both contained within the same result display, marking each result appropriately as Certain or Uncertain. Uncertain words, which were sometimes non-existent should be able to find at least some place (sometimes) in your top scored results. The program as a whole should also spend less time loading (as it's only creating one tree, which is more efficient than making two) and less time to return results. Processing results should also look better, as it doesn't need to run the same recursive function twice, and therefore display the progress of checking each tile twice. There were several relatively minor changes made to some functions in order to work with the new system, as well as other minor optimizations to some core functions. Also added in more commenting and tried to tidy up some code. Lastly removed the no-longer-necessary enable1.txt and enable2.txt, to make way for the modified enable.txt.

This is a test release, as I haven't yet tested the code very much, so there may be bugs still in there. I don't think there will be or if there are any they're too major, as the change to the new system was relatively straight-forward (due to some functions already being unintentionally very versatile and as such well suited to change) so not too many things were absolutely required to change. As always, though, will continue to search for bugs.

And as a final note, I have drawn up some plans for adding in a pygame-based GUI, so that is in the works. I get tired of manually editing the board.txt file every time a change is made, and every time I find a new board to work with. Planning on making it easier to input new boards, see/sort results, view results in relation to the board, and increase the total usefulness of the entire program. I'm not actively developing this as much as I used to, so it's not going to happen for a while, but it's something I already have planned out and will start implementing.